Major debug draw fix

+++ BITMASK/DrawDebug/BoxDebugrender.h
@@ -35,7 +35,7 @@ class DebugDraw : public b2Draw
 		void DrawPoint(const b2Vec2& p, float32 size, const b2Color& color);
 		void DrawString(int x, int y, const char* string);
 		void DrawAABB(b2AABB* aabb, const b2Color& color);
-		sf::Color B2SFColor(const b2Color &color, int alpha);
+		sf::Color B2SFColor(const b2Color &color, float alpha);
 
 	private:
 		sf::RenderWindow *window;
